next Steps from old project

+++ app/Http/Controllers/UserController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\Role;
+use App\Buisness\Enum\PermissionEnum;
 use App\User;
 use App\Http\Requests\UserRequest;
 use Illuminate\Support\Facades\Hash;
@@ -11,9 +11,9 @@ class UserController extends Controller
 {
     public function __construct()
     {
-        $this->authorizeResource(User::class);
+        //FIXME: 403 - Forbidden
+        //$this->authorizeResource(User::class);
     }
-
     /**
      * Display a listing of the users
      *
@@ -22,20 +22,19 @@ class UserController extends Controller
      */
     public function index(User $model)
     {
-        $this->authorize('manage-users', User::class);
-
-        return view('users.index', ['users' => $model->with('role')->get()]);
+        $this->authorize(PermissionEnum::userManagement()->getValue(), User::class);
+        return view('users.index', ['users' => User::paginate(25)]);
     }
 
     /**
      * Show the form for creating a new user
      *
-     * @param  \App\Role  $model
      * @return \Illuminate\View\View
      */
-    public function create(Role $model)
+    public function create()
     {
-        return view('users.create', ['roles' => $model->get(['id', 'name'])]);
+        $this->authorize(PermissionEnum::userManagement()->getValue(), User::class);
+        return view('users.create');
     }
 
     /**
@@ -47,24 +46,21 @@ class UserController extends Controller
      */
     public function store(UserRequest $request, User $model)
     {
-        $model->create($request->merge([
-            'picture' => $request->photo ? $request->photo->store('profile', 'public') : null,
-            'password' => Hash::make($request->get('password'))
-        ])->all());
-
-        return redirect()->route('user.index')->withStatus(__('User successfully created.'));
+        $this->authorize(PermissionEnum::userManagement()->getValue(), User::class);
+        $model->create($request->merge(['password' => Hash::make($request->get('password'))])->all());
+        return redirect()->route('user.index')->withStatus(__('Teammitglied erfolgreich erstellt.'));
     }
 
     /**
      * Show the form for editing the specified user
      *
      * @param  \App\User  $user
-     * @param  \App\Role  $model
      * @return \Illuminate\View\View
      */
-    public function edit(User $user, Role $model)
+    public function edit(User $user)
     {
-        return view('users.edit', ['user' => $user->load('role'), 'roles' => $model->get(['id', 'name'])]);
+        $this->authorize(PermissionEnum::userManagement()->getValue(), User::class);
+        return view('users.edit', compact('user'));
     }
 
     /**
@@ -76,14 +72,13 @@ class UserController extends Controller
      */
     public function update(UserRequest $request, User $user)
     {
+        $this->authorize(PermissionEnum::userManagement()->getValue(), User::class);
         $user->update(
-            $request->merge([
-                'picture' => $request->photo ? $request->photo->store('profile', 'public') : $user->picture,
-                'password' => Hash::make($request->get('password'))
-            ])->except([$request->get('password') ? '' : 'password'])
-        );
+            $request->merge(['password' => Hash::make($request->get('password'))])
+                ->except([$request->get('password') ? '' : 'password']
+                ));
 
-        return redirect()->route('user.index')->withStatus(__('User successfully updated.'));
+        return redirect()->route('user.index')->withStatus(__('Teammitglied erfolgreich geändert.'));
     }
 
     /**
@@ -94,8 +89,8 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        $this->authorize(PermissionEnum::userManagement()->getValue(), User::class);
         $user->delete();
-
-        return redirect()->route('user.index')->withStatus(__('User successfully deleted.'));
+        return redirect()->route('user.index')->withStatus(__('Teammitglied erfolgreich gelöscht'));
     }
 }
